[SofaSimulationCore] Move MechanicalGetDimensionVisitor to dedicated folder
